pyproject.toml
[project]
name = "strawb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "granian>=2.3.1",
    "strawberry-graphql[channels]>=0.270.1",
]

[dependency-groups]
dev = [
    "daphne>=4.2.0",
    "django-stubs>=5.2.0",
]
